Fix bad printf

+++ src/workload.cpp
@@ -93,7 +93,8 @@ void Workload::register_smpi_applications()
             SmpiProfileData * data = (SmpiProfileData *) profile->data;
 
             string job_id_str = name + "!" + to_string(job->number);
-            XBT_INFO("Registering app. instance='%s', nb_process=%d", job_id_str.c_str(), data->trace_filenames.size());
+            XBT_INFO("Registering app. instance='%s', nb_process=%d",
+                     job_id_str.c_str(), (int) data->trace_filenames.size());
             SMPI_app_instance_register(job_id_str.c_str(), smpi_replay_process, data->trace_filenames.size());
         }
     }
